fix(aimi): AwakeRestingHeartRate day key is a Long epoch day
kotlinx-datetime 0.8 returns Long from toEpochDays. Drop the unused
AwakeRestingHeartRate import on the Android plugin.

+++ plugins/aps/src/commonMain/kotlin/app/aaps/plugins/aps/openAPSAIMI/physio/AwakeRestingHeartRate.kt
@@ -111,6 +111,6 @@ object AwakeRestingHeartRate {
         return hour in AWAKE_FIRST_HOUR until NIGHT_FIRST_HOUR
     }
 
-    private fun localDay(timestampMs: Long, zoneId: TimeZone): Int =
+    private fun localDay(timestampMs: Long, zoneId: TimeZone): Long =
         Instant.fromEpochMilliseconds(timestampMs).toLocalDateTime(zoneId).date.toEpochDays()
 }
